requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a1634+g7f4c57d
idna==2.8
multidict==4.5.2
pycparser==2.19
PyMySQL==0.9.3
six==1.12.0
SQLAlchemy==1.2.17
websockets==6.0
yarl==1.3.0
